pyproject.toml
[project]
name = "case-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.8.3",
    "chromadb>=1.2.0",
    "huggingface-hub>=0.35.3",
    "langchain>=1.0.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4",
    "langchain-huggingface>=1.0.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.14.1",
    "sentence-transformers>=5.1.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]
